Implement the ERC721 interface, including the optional Enumerable and Metadata.
Since the Constitution already supported most of the functionality from the specification, the goal for this initial implementation was to leave existing code as untouched as possible.

Some additions to Ships were necessary to support all ERC721 operations fully. The addition of simple getters for existing data indicates we may want to consider implementing those for all ship data, to ensure it offers all functionality future Constitutions may need.
An `isOperator(uint32, address)` was not implemented because it hits a compiler bug that causes it to be seen as identical to the `isOperator(address, address)`. Luckily it is not strictly necessary to have.
The existing `transferrer` property fills the role of ERC721's `approved`. `operators` are an addition. They still only receive transfer permissions, and can't start, rekey, escape etc. any ships. Such actions aren't in line with the intended use of "consignment to third party brokers/wallets/auctioneers" as described in ERC721.

The `tokenURI(uint256)` implementation is merely a suggestion, but might be nice to fully support, especially when ships get "identicons". The other metadata fields, `name` and `symbol`, are of course also still up for change.

Compiles (also see Constitution.sol:12), but as of yet untested.