Make the spawning of ships follow a withdraw pattern. Unless the ship is spawned to the caller (in which case spawn will function as it used to), the spawned ship will get its transferProxy set to the recipient. To make cancelling of this transfer possible, the ship's owner is set to match its parent's.
This was done to prevent loss off address space. Clients are advised to use setTransferProxy() over transferShip() (and their ERC721 equivalents) whenever possible.

This means that inactive ships can have owners. Most functions' modifiers have changed to check for both ownership and ship activity. The only operations that can be performed with inactive ships are transfer-related operations.

Most contracts are fine to continue operating as they did, since they implement some variant of withdraw pattern themselves.
The PlanetSale contract has changed to spawn to itself and then immediately transfer to the recipient, to avoid giving the planet's parent the possibility of interfering with the transfer.