Restructure ownership and permissions
The desired change is having managers and delegates as proxies per ship,
rather than per owner.
Adding more fields to the Hull struct pushed us over the limit of 16
fields per struct, prompting the slightly broader change you find in
this commit.

We chose to split off the ownership and permission management data from
other ship state. This makes a clearer distinction between the things
that matter to Urbit and the things that matter on-chain.