ecliptic: prepare for naive rollup
Define special `depositAddress`.  When we transfer to this address, we
clear all proxies so that the ship can no longer be controlled here on
L1.

If we set the spawn proxy to this address, then we disallow spawning on
L1.  We also disallow changing the spawn proxy any more, since we can't
know which ships were already spawned on L2.

During escape, both ships must be on L1 to make the change on L1.  If
either are on L2, the escape must happen on L2.  If both ships are on
L1, the escape can happen on either layer.